Sourced: move the ID-specific logic onto the parent class
Several of the methods/formatters defined on Versioned can also be
defined on Sourced, and this makes them accessible to classes that
extend S but not V.
